refactor: modal now wraps its contents with ModalContainer
